YP-37: fix
