+++ packages/client/src/Components/App/App.tsx
@@ -4,12 +4,7 @@ import { Outlet } from 'react-router-dom';
 
 import { DebugPanel, Menu } from '@Components';
 import { useRender, useUpdate } from '@Game';
-import {
-  useGameLoop,
-  useGetCanvasSize,
-  useSetCanvasContext,
-  useToggleFullscreen,
-} from '@Hooks';
+import { useGameLoop, useGetCanvasSize, useSetCanvasContext } from '@Hooks';
 
 import './App.scss';
 
@@ -37,10 +32,24 @@ export const App = () => {
     };
   }, [ctx]);
 
+  const toggleFullscreen = () => {
+    const elem = document.querySelector('.game-field__canvas');
+
+    if (!document.fullscreenElement) {
+      elem?.requestFullscreen().catch((err) => {
+        alert(
+          `Error attempting to enable fullscreen mode: ${err.message} (${err.name})`,
+        );
+      });
+    } else {
+      document.exitFullscreen();
+    }
+  };
+
   useEffect(() => {
     function handleKeyDown(event: KeyboardEvent) {
       if (event.shiftKey && event.key === 'F11') {
-        useToggleFullscreen();
+        toggleFullscreen();
       }
     }
     document.addEventListener('keydown', handleKeyDown);
@@ -63,7 +72,11 @@ export const App = () => {
       <Menu />
 
       <section className={'game-field'}>
-        <DebugPanel onStart={startGame} onStop={stopGame} />
+        <DebugPanel
+          onStart={startGame}
+          onStop={stopGame}
+          toggleFullscreen={toggleFullscreen}
+        />
 
         <div className={'game-field__canvas-container'}>
           <canvas
