+++ packages/client/src/Components/DebugPanel/DebugPanel.props.ts
@@ -1,4 +1,5 @@
 export interface IDebugPanelProps {
   onStart: () => void;
   onStop: () => void;
+  toggleFullscreen: () => void;
 }
