+++ packages/client/src/Hooks/index.ts
@@ -1,4 +1,3 @@
 export { useGameLoop } from './UseGameLoop';
 export { useGetCanvasSize } from './UseGetCanvasSize';
 export { useSetCanvasContext } from './UseSetCanvasContext';
-export { useToggleFullscreen } from './UseToggleFullscreen';
